+++ loki/visitors/ir_graph.py
