+++ tests/test_ir_graph.py
@@ -10,7 +10,7 @@ from pathlib import Path
 import pytest
 from conftest import graphviz_present
 from loki import Sourcefile
-from loki.visitors.pretty_visualize import ir_graph, GraphCollector
+from loki.visitors.ir_graph import ir_graph, GraphCollector
 
 
 @pytest.fixture(scope="module", name="here")
